+++ src/sdk/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added support for adding, updating and removing audiences on navigation nodes #1065 [jansenbe - Bert Jansen]
 - Admin library: added 43 new properties to `ITenantProperties` [jansenbe - Bert Jansen]
 - Admin library: added 15 new properties to `ISiteCollectionProperties` [jansenbe - Bert Jansen]
+- Bulk user validation method `ValidateUsersAsync` and `ValidateAndEnsureUsersAsync` on `IWeb` [jansenbe - Bert Jansen]
 
 ### Changed